triedb/pathdb: fix windows build issue
